Add a ‘Desync’ class as the main interface for using the async queues
The actual async methods are a bit unweildy to call directly as Rust will insist that things be cloned all over the place. This uses some knowledge of the behaviour of the queue (and some unsafe code) to wrap a data structure such that it can be accessed asynchronously or synchronously.