Pass the waker through instead of the context referencing it
This seems to have broken in the latest release of Rust futures as Context used to be 'Send' so we were previously OK with wrapping it up beforehand. Not sure if the old code was really 'valid' or not, but it does make sense to create multiple contexts when repeatedly polling the stream to drain it.